requirements.txt
async-timeout==4.0.3
backtrader==1.9.78.123
binance-futures-connector==4.1.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
greenlet==3.1.1
idna==3.10
importlib-metadata==6.7.0
numpy==1.21.6
pandas==1.3.5
pycryptodome==3.21.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.2
redis==5.0.8
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.7.1
urllib3==1.26.7
websocket-client==1.6.1
zipp==3.15.0
